create search shared component
