+++ src/pages/StoresByCity.vue
@@ -1,8 +1,11 @@
 <template>
   <div class="">
+    <SearchBlock @onSearch="onSearch($event)" @onClear="onClear()"></SearchBlock>
     <div class="block" v-if="!this.dataStore.isLoading">
-      <CityList :list="getCitiesName" isNotFound="false" />
-      <TableStores :item="getCitiesName[this.cityStore.activeIndex]" />
+      <CityList :list="filteredCities ? filteredCities : getUniqCities" :isNotFound="isNotFound" />
+      <TableStores
+        :item="filteredCities ? filteredCities[this.cityStore.activeIndex] : getUniqCities[this.cityStore.activeIndex]"
+      />
     </div>
     <div v-else>Loading...</div>
   </div>
@@ -11,25 +14,33 @@
 <script>
 import CityList from '@/components/CityList';
 import TableStores from '@/components/TableStores';
-import { mapState } from 'vuex';
+import SearchBlock from '@/components/SearchBlock';
+import { mapGetters, mapState } from 'vuex';
 export default {
   components: {
     CityList,
     TableStores,
+    SearchBlock,
   },
   data() {
     return {
-      filteredCities: [],
+      filteredCities: undefined,
     };
   },
   mounted() {
     this.$store.dispatch('getData');
   },
-  methods: {},
+  methods: {
+    onSearch(str) {
+      this.filteredCities = this.getCitySearchResult(str);
+    },
+    onClear() {},
+  },
   computed: {
     ...mapState(['dataStore', 'cityStore']),
-    getCitiesName() {
-      return this.$store.getters.getUniqCities;
+    ...mapGetters(['getUniqCities', 'getCitySearchResult']),
+    isNotFound() {
+      return this.filteredCities && this.filteredCities.length === 0;
     },
   },
 };
